add support for R lists
This is a significant addition. I realized that for Rgo/sexp to really be useful, passing lists or multiple returns was necessary.

This adds a type to sexp, called a List, which is just a slice of GoSEXP. This is converted to an R list using the List2sexp function. Like the other 2sexp functions, it is itself basically a wrapper for some quick and dirty C code that allocates elements within the list.

The List type exists to make sure that users of sexp only use lists when they really mean to, and it forces them to consider what actually belongs in the list they're returning an not. I think this is more robust than using something like a variadic funtion to insert several SEXP/GoSEXP objects into a list.