better matrix implementation
This is more performant and carries less overhead. I don't know why I didn't think of this first.

Basically, this implementation is based on a matrix struct that contains two int fields that describe the matrix dimensions and then a vector containing the data. The vector is each row of the matrix, one after another. It's similar to the vectorized matrix that is passed between R and Go, except it drops the first element that contains the number of columns.