Fixing bug where AsStrings() didn't handle strings with more than about 20 characters correctly.
I was never able to identify a clear pattern to the behaviour, nor was I able to identify quite why the string conversion went awry for different strings, but I found that the C extraction functions were correctly streaming the bytes from R to Go. Things were either going wrong due to my misunderstanding of how to correctly use C.GoSring and/or the += operator.

Whatever the cause, I was able to fix the issue by putting the stream of bytes from R into a Go byte slice, and then converting it to a string once all the bytes from R have been read.