Fixing a tricky bug in which the NewGoSEXP function would lead to a runtime crash (not a panic!) if the user provided a SEXP instead of a *SEXP. The issue is that the Go type checking is not good enough (because what we're doing in rsexp isn't explicitly supported by Go, nor should it be) to be able to tell the difference between a SEXP and *SEXP well enough to avoid a fatal runtime error (see https://github.com/golang/go/issues/28460).
The answer, therefore, is to either have the user pass in just a SEXP, or to support both SEXP and *SEXP as inputs. I chose the latter to keep backwards compatibility.

I updated the demonstration shiny app accordingly, and provided comments in NewGoSEXP to make the logic of the function easier to follow.